+++ kimchi/src/tests/foreign_field_add.rs
@@ -471,7 +471,7 @@ fn test_zero_sum_native() {
     );
 
     // Check result is the native modulus
-    let native_limbs = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(native_modulus);
+    let native_limbs = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&native_modulus);
     check_result(witness, vec![native_limbs]);
 }
 
@@ -501,7 +501,7 @@ fn test_max_number() {
     // compute result in the foreign field after taking care of the exceeding bits
     let sum = secp256k1_max() + secp256k1_max();
     let sum_mod = sum - secp256k1_modulus();
-    let sum_mod_limbs = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(sum_mod);
+    let sum_mod_limbs = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&sum_mod);
     check_ovf(witness.clone(), PallasField::one());
     check_result(witness, vec![sum_mod_limbs]);
 }
@@ -513,11 +513,11 @@ fn test_max_number() {
 // and it is checked that in both cases the result is the same
 fn test_zero_minus_one() {
     // FIRST AS NEG
-    let right_be_neg = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(One::one())
+    let right_be_neg = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&One::one())
         .neg(&secp256k1_modulus())
         .to_biguint();
     let right_for_neg: ForeignElement<PallasField, LIMB_BITS, 3> =
-        ForeignElement::from_biguint(right_be_neg.clone());
+        ForeignElement::from_biguint(&right_be_neg);
     let (witness_neg, _index) = test_ffadd(
         secp256k1_modulus(),
         vec![BigUint::zero(), right_be_neg],
@@ -540,7 +540,7 @@ fn test_zero_minus_one() {
 // test 1 - 1 + 1 where (-1) is in the foreign field
 // the first check is done with sub(1, 1) and then with add(neg(neg(1)))
 fn test_one_minus_one_plus_one() {
-    let neg_neg_one = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(One::one())
+    let neg_neg_one = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&One::one())
         .neg(&secp256k1_modulus())
         .neg(&secp256k1_modulus())
         .to_biguint();
@@ -556,7 +556,7 @@ fn test_one_minus_one_plus_one() {
         witness,
         vec![
             ForeignElement::zero(),
-            ForeignElement::from_biguint(One::one()),
+            ForeignElement::from_biguint(&One::one()),
         ],
     );
 }
@@ -567,12 +567,12 @@ fn test_one_minus_one_plus_one() {
 // then tested as 0 - 1 - 1 )
 // TODO: tested as 0 - ( 1 + 1) -> put sign in front of left instead (perhaps in the future we want this)
 fn test_minus_minus() {
-    let neg_one_for = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(One::one())
+    let neg_one_for = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&One::one())
         .neg(&secp256k1_modulus());
     let neg_one = neg_one_for.to_biguint();
     let two = BigUint::from(2u32);
     let neg_two =
-        ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(two).neg(&secp256k1_modulus());
+        ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&two).neg(&secp256k1_modulus());
     let (witness_neg, _index) = test_ffadd(
         secp256k1_modulus(),
         vec![neg_one.clone(), neg_one],
@@ -820,7 +820,7 @@ fn test_zero_sub_fmax() {
         &[FFOps::Sub],
         false,
     );
-    let negated = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(secp256k1_max())
+    let negated = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&secp256k1_max())
         .neg(&secp256k1_modulus());
     check_result(witness, vec![negated]);
 }
@@ -840,7 +840,7 @@ fn test_pasta_add_max_vesta() {
         false,
     );
     let right = right_input % vesta_modulus;
-    let right_foreign = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(right);
+    let right_foreign = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&right);
     check_result(witness, vec![right_foreign]);
 }
 
@@ -856,7 +856,7 @@ fn test_pasta_sub_max_vesta() {
         false,
     );
     let neg_max_vesta =
-        ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(right_input).neg(&vesta_modulus);
+        ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&right_input).neg(&vesta_modulus);
     check_result(witness, vec![neg_max_vesta]);
 }
 
@@ -872,7 +872,7 @@ fn test_pasta_add_max_pallas() {
         false,
     );
     let right = right_input % vesta_modulus;
-    let foreign_right = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(right);
+    let foreign_right = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&right);
     check_result(witness, vec![foreign_right]);
 }
 
@@ -888,7 +888,7 @@ fn test_pasta_sub_max_pallas() {
         false,
     );
     let neg_max_pallas =
-        ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(right_input).neg(&vesta_modulus);
+        ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&right_input).neg(&vesta_modulus);
     check_result(witness, vec![neg_max_pallas]);
 }
 
@@ -908,7 +908,7 @@ fn test_random_add() {
         false,
     );
     let result = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(
-        (left_big + right_big) % foreign_mod,
+        &((left_big + right_big) % foreign_mod),
     );
     check_result(witness, vec![result]);
 }
@@ -930,10 +930,10 @@ fn test_random_sub() {
     );
     let result = if left_big < right_big {
         ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(
-            left_big + foreign_mod - right_big,
+            &(left_big + foreign_mod - right_big),
         )
     } else {
-        ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(left_big - right_big)
+        ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&(left_big - right_big))
     };
     check_result(witness, vec![result]);
 }
@@ -956,7 +956,7 @@ fn test_foreign_is_native_add() {
     );
     // check result was computed correctly
     let sum_big = compute_sum(pallas, &left_input, &right_input);
-    let result = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(sum_big.clone());
+    let result = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&sum_big);
     check_result(witness, vec![result.clone()]);
     // check result is in the native field
     let two_to_limb = PallasField::from(TWO_TO_LIMB);
@@ -989,7 +989,7 @@ fn test_foreign_is_native_sub() {
     );
     // check result was computed correctly
     let dif_big = compute_dif(pallas, &left_input, &right_input);
-    let result = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(dif_big.clone());
+    let result = ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&dif_big);
     check_result(witness, vec![result.clone()]);
     // check result is in the native field
     let two_to_limb = PallasField::from(TWO_TO_LIMB);
@@ -1027,7 +1027,7 @@ fn test_random_small_add() {
     check_result(
         witness,
         vec![ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(
-            result,
+            &result,
         )],
     );
 }
@@ -1056,7 +1056,7 @@ fn test_random_small_sub() {
     check_result(
         witness,
         vec![ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(
-            result,
+            &result,
         )],
     );
 }
@@ -1250,7 +1250,7 @@ fn test_random_chain() {
                 FFOps::Sub => compute_dif(foreign_mod.clone(), &left[i], &inputs[i + 1]),
             };
             left.push(result.to_bytes_be());
-            ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(result)
+            ForeignElement::<PallasField, LIMB_BITS, 3>::from_biguint(&result)
         })
         .collect();
     check_result(witness, results);